Jst fixed a few stuff

+++ KartBookingSession/Model/tblCoachInfo.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Numerics;
 
 namespace KartBookingSession.Model
 {
@@ -11,10 +12,10 @@ namespace KartBookingSession.Model
         public int coachinfo_id { get; set; }
         public int coach_id { get; set; }
         public string email { get; set; }
-        public int phone_number { get; set; }
+        public BigInteger phone_number { get; set; }
         public string experience_lvl { get; set; }
 
-        public CoachInfo(int CoachInfoID, int CoachID, string Email, int PhoneNumber, string ExperienceLvl)
+        public CoachInfo(int CoachInfoID, int CoachID, string Email, BigInteger PhoneNumber, string ExperienceLvl)
         {
             coachinfo_id = CoachInfoID;
             coach_id = CoachID;
